# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "The Grand Azure - Hotel Management System (HMS) API",
    "description": "Enterprise-grade, domain-modular Hotel Management System REST API specification adhering to HFTP Uniform System of Accounts for the Lodging Industry (USALI).",
    "version": "1.0.0",
    "contact": {
      "name": "The Grand Azure Engineering",
      "email": "dev@grandazure.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local Development Server"
    },
    {
      "url": "https://api.grandazure.com/api",
      "description": "Cloud Production Gateway"
    }
  ],
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["Identity & RBAC"],
        "summary": "Authenticate system user credentials",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string", "example": "admin" },
                  "password": { "type": "string", "example": "admin123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Authentication successful" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/users": {
      "post": {
        "tags": ["Identity & RBAC"],
        "summary": "Provision new staff account with designated role",
        "responses": {
          "201": { "description": "User created successfully" },
          "400": { "description": "Username already exists or invalid role" }
        }
      }
    },
...
  }
}